Refactor endpoints into FastAPI routers
