+++ routers/users.py
@@ -1,0 +1,43 @@
+from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordRequestForm
+from sqlalchemy.orm import Session
+
+from database import get_db
+from models import User as UserModel
+from schemas.user import UserCreate, UserResponse, Token
+from main import (
+    get_password_hash,
+    verify_password,
+    create_access_token,
+    get_current_user,
+)
+
+router = APIRouter()
+
+@router.post("/users/", response_model=UserResponse)
+def create_user(user: UserCreate, db: Session = Depends(get_db)):
+    existing_user = db.query(UserModel).filter_by(username=user.username).first()
+    if existing_user:
+        raise HTTPException(status_code=400, detail="Username already registered")
+    db_user = UserModel(
+        username=user.username, hashed_password=get_password_hash(user.password)
+    )
+    db.add(db_user)
+    db.commit()
+    db.refresh(db_user)
+    return db_user
+
+@router.post("/token", response_model=Token)
+def login_for_access_token(
+    form_data: OAuth2PasswordRequestForm = Depends(),
+    db: Session = Depends(get_db),
+):
+    user = db.query(UserModel).filter_by(username=form_data.username).first()
+    if not user or not verify_password(form_data.password, user.hashed_password):
+        raise HTTPException(status_code=400, detail="Incorrect username or password")
+    access_token = create_access_token(data={"sub": user.username})
+    return {"access_token": access_token, "token_type": "bearer"}
+
+@router.get("/users/me", response_model=UserResponse)
+def read_users_me(current_user: UserModel = Depends(get_current_user)):
+    return current_user
